// src/app/modules/cxc/model/nota-debito-emitir.ts
import { Facturador } from './facturador';
import { Titular } from '../../tsr/model/titular';
import { Usuario } from '../../../shared/model/usuario';
import { PuntoEmision } from './puntos-emision';
import { FacturaEmitir } from './factura-emitir';

export interface NotaDebitoEmitir {
  id: number;
  tipoComprobante: string;
  facturador: Facturador;
  titular: Titular;
  tipoDoc: string;
  numero: string;
  numEstablecimiento: string;
  numPtoEmision: string;
  secuencial: string;
  ambiente: number;
  clave: string;
  fecha: Date;
  tipoDocModificado: string;
  numDocModificado: string;
  fechaEmisionDM: Date;
  observacion: string;
  subtotal: number;
  subcero: number;
  pIVA: number;
  vIVA: number;
  vICE: number;
  vIRBPNR: number;
  descuento: number;
  porDescuento: number;
  propina: number;
  subsidio: number;
  total: number;
  ptoEmision: PuntoEmision;
  usuario: Usuario;
  pathGen: string;
  autorizacion: string;
  fechaAutorizacion: string;
  estado: number;
  estadoEmision: number;
  motivoAnulacion?: string | null;
  fechaAnulacion?: string | null;
  usuarioAnulacion?: string | null;
  factura?: FacturaEmitir;
}